village presque terminé

+++ RoyalRevolt/Macro.h
@@ -60,8 +60,8 @@ static Vector2f GetRandomPosition()
 
 static bool IsSameName(const string& _name1, const string& _name2)
 {
-	int _size1 = _name1.size();
-	int _size2 = _name2.size();
+	int _size1 = static_cast<int>(_name1.size());
+	int _size2 = static_cast<int>(_name2.size());
 	int _size = 0;
 	bool _isSame = true;
